+++ tests/integration/high_availability/test_upgrade.py
@@ -3,223 +3,217 @@
 
 import json
 import logging
-import pathlib
 import shutil
 import zipfile
 from contextlib import suppress
 from pathlib import Path
-from time import sleep
-from typing import Union
 
+import jubilant_backports
 import pytest
-from pytest_operator.plugin import OpsTest
+from jubilant_backports import Juju, TaskError
 
-from .. import juju_
-from ..helpers import (
-    get_leader_unit,
-    get_primary_unit,
+from .high_availability_helpers_new import (
+    CHARM_METADATA,
+    check_mysql_units_writes_increment,
+    get_app_leader,
+    get_app_units,
+    get_k8s_stateful_set_partitions,
+    get_mysql_primary_unit,
+    get_mysql_variable_value,
     get_relation_data,
     get_unit_by_number,
-    retrieve_database_variable_value,
+    wait_for_apps_status,
+    wait_for_unit_message,
+    wait_for_unit_status,
 )
-from .high_availability_helpers import (
-    METADATA,
-    ensure_all_units_continuous_writes_incrementing,
-    get_sts_partition,
-    relate_mysql_and_application,
-)
-
-logger = logging.getLogger(__name__)
-
-TIMEOUT = 15 * 60
 
 MYSQL_APP_NAME = "mysql-k8s"
-TEST_APP_NAME = "mysql-test-app"
+MYSQL_TEST_APP_NAME = "mysql-test-app"
+
+MINUTE_SECS = 60
+
+logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 
 @pytest.mark.abort_on_fail
-async def test_deploy_latest(ops_test: OpsTest) -> None:
-    """Simple test to ensure that the mysql and application charms get deployed."""
-    await ops_test.model.deploy(
-        MYSQL_APP_NAME,
-        application_name=MYSQL_APP_NAME,
-        num_units=3,
+def test_deploy_latest(juju: Juju) -> None:
+    """Simple test to ensure that the MySQL and application charms get deployed."""
+    logging.info("Deploying MySQL cluster")
+    juju.deploy(
+        charm=MYSQL_APP_NAME,
+        app=MYSQL_APP_NAME,
+        base="ubuntu@22.04",
         channel="8.0/edge",
-        trust=True,
         config={"profile": "testing"},
-        base="ubuntu@22.04",
+        num_units=3,
+        trust=True,
     )
-    await ops_test.model.deploy(
-        TEST_APP_NAME,
-        application_name=TEST_APP_NAME,
-        num_units=1,
+    juju.deploy(
+        charm=MYSQL_TEST_APP_NAME,
+        app=MYSQL_TEST_APP_NAME,
+        base="ubuntu@22.04",
         channel="latest/edge",
-        base="ubuntu@22.04",
+        num_units=1,
+        trust=False,
     )
 
-    await relate_mysql_and_application(ops_test, MYSQL_APP_NAME, TEST_APP_NAME)
-    logger.info("Wait for applications to become active")
-    await ops_test.model.wait_for_idle(
-        apps=[MYSQL_APP_NAME, TEST_APP_NAME],
-        status="active",
-        timeout=TIMEOUT,
-        raise_on_error=False,
+    juju.integrate(
+        f"{MYSQL_APP_NAME}:database",
+        f"{MYSQL_TEST_APP_NAME}:database",
     )
-    assert len(ops_test.model.applications[MYSQL_APP_NAME].units) == 3
+
+    logging.info("Wait for applications to become active")
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, MYSQL_APP_NAME, MYSQL_TEST_APP_NAME
+        ),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
 
 
 @pytest.mark.abort_on_fail
-async def test_pre_upgrade_check(ops_test: OpsTest) -> None:
+def test_pre_upgrade_check(juju: Juju) -> None:
     """Test that the pre-upgrade-check action runs successfully."""
-    mysql_units = ops_test.model.applications[MYSQL_APP_NAME].units
+    mysql_leader = get_app_leader(juju, MYSQL_APP_NAME)
+    mysql_units = get_app_units(juju, MYSQL_APP_NAME)
 
-    logger.info("Get leader unit")
-    leader_unit = await get_leader_unit(ops_test, MYSQL_APP_NAME)
-    assert leader_unit is not None, "No leader unit found"
+    logging.info("Run pre-upgrade-check action")
+    task = juju.run(unit=mysql_leader, action="pre-upgrade-check")
+    task.raise_on_failure()
 
-    logger.info("Run pre-upgrade-check action")
-    await juju_.run_action(leader_unit, "pre-upgrade-check")
+    logging.info("Assert slow shutdown is enabled")
+    for unit_name in mysql_units:
+        value = get_mysql_variable_value(juju, MYSQL_APP_NAME, unit_name, "innodb_fast_shutdown")
+        assert value == 0
 
-    logger.info("Assert slow shutdown is enabled")
-    for unit in mysql_units:
-        value = await retrieve_database_variable_value(ops_test, unit, "innodb_fast_shutdown")
-        assert value == 0, f"innodb_fast_shutdown not 0 at {unit.name}"
+    logging.info("Assert primary is set to leader")
+    mysql_primary = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    assert mysql_primary == f"{MYSQL_APP_NAME}/0", "Primary unit not set to unit 0"
 
-    primary_unit = await get_primary_unit(ops_test, leader_unit, MYSQL_APP_NAME)
-
-    assert primary_unit.name == f"{MYSQL_APP_NAME}/0", "Primary unit not set to unit 0"
-
-    logger.info("Assert partition is set to 2")
-
-    assert get_sts_partition(ops_test, MYSQL_APP_NAME) == 2, "Partition not set to 2"
+    logging.info("Assert partition is set to 2")
+    assert get_k8s_stateful_set_partitions(juju, MYSQL_APP_NAME) == 2, "Partition not set to 2"
 
 
 @pytest.mark.abort_on_fail
-async def test_upgrade_from_edge(ops_test: OpsTest, charm, continuous_writes, credentials) -> None:
-    logger.info("Ensure continuous_writes")
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
+def test_upgrade_from_edge(juju: Juju, charm: str, continuous_writes) -> None:
+    """Update the second cluster."""
+    logging.info("Ensure continuous writes are incrementing")
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
-    resources = {"mysql-image": METADATA["resources"]["mysql-image"]["upstream-source"]}
-    application = ops_test.model.applications[MYSQL_APP_NAME]
-
-    logger.info("Refresh the charm")
-    await application.refresh(path=charm, resources=resources)
-
-    logger.info("Wait for upgrade to complete on first upgrading unit")
-    # highest ordinal unit always the first to upgrade
-    unit = get_unit_by_number(MYSQL_APP_NAME, application.units, 2)
-
-    await ops_test.model.block_until(
-        lambda: unit.workload_status_message == "upgrade completed", timeout=TIMEOUT
+    logging.info("Refresh the charm")
+    juju.refresh(
+        app=MYSQL_APP_NAME,
+        path=charm,
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
     )
 
-    leader_unit = await get_leader_unit(ops_test, MYSQL_APP_NAME)
-    assert leader_unit is not None, "No leader unit found"
+    mysql_app_leader = get_app_leader(juju, MYSQL_APP_NAME)
+    mysql_upgrade_unit = get_unit_by_number(juju, MYSQL_APP_NAME, 2)
 
-    logger.info("Resume upgrade")
-    while get_sts_partition(ops_test, MYSQL_APP_NAME) == 2:
-        # resume action sometime fails in CI, no clear reason
-
-        with suppress(AssertionError):
-            # ignore action return error as it is expected when
-            # the leader unit is the next one to be upgraded
-            # due it being immediately rolled when the partition
-            # is patched in the statefulset
-            await juju_.run_action(leader_unit, "resume-upgrade")
-
-    logger.info("Wait for upgrade to complete")
-    await ops_test.model.block_until(
-        lambda: all(unit.workload_status == "active" for unit in application.units),
-        timeout=TIMEOUT,
+    logging.info("Wait for upgrade to complete on first upgrading unit")
+    juju.wait(
+        ready=wait_for_unit_message(MYSQL_APP_NAME, mysql_upgrade_unit, "upgrade completed"),
+        timeout=10 * MINUTE_SECS,
     )
 
-    logger.info("Ensure continuous_writes")
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
+    logging.info("Resume upgrade")
+    while get_k8s_stateful_set_partitions(juju, MYSQL_APP_NAME) == 2:
+        # ignore action return error as it is expected when
+        # the leader unit is the next one to be upgraded
+        # due it being immediately rolled when the partition
+        # is patched in the stateful set
+        with suppress(TaskError):
+            task = juju.run(unit=mysql_app_leader, action="resume-upgrade")
+            task.raise_on_failure()
+
+    logging.info("Wait for upgrade to complete")
+    juju.wait(
+        ready=lambda status: jubilant_backports.all_active(status, MYSQL_APP_NAME),
+        timeout=20 * MINUTE_SECS,
+    )
+
+    logging.info("Ensure continuous writes are incrementing")
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
 
 @pytest.mark.abort_on_fail
-async def test_fail_and_rollback(ops_test, charm, continuous_writes, credentials) -> None:
-    logger.info("Get leader unit")
-    leader_unit = await get_leader_unit(ops_test, MYSQL_APP_NAME)
-    assert leader_unit is not None, "No leader unit found"
+def test_fail_and_rollback(juju: Juju, charm: str, continuous_writes) -> None:
+    """Test an upgrade failure and its rollback."""
+    mysql_app_leader = get_app_leader(juju, MYSQL_APP_NAME)
+    mysql_app_units = get_app_units(juju, MYSQL_APP_NAME)
+    mysql_upgrade_unit = get_unit_by_number(juju, MYSQL_APP_NAME, 2)
 
-    logger.info("Run pre-upgrade-check action")
-    await juju_.run_action(leader_unit, "pre-upgrade-check")
+    logging.info("Run pre-upgrade-check action")
+    task = juju.run(unit=mysql_app_leader, action="pre-upgrade-check")
+    task.raise_on_failure()
 
-    fault_charm = Path("/tmp/", pathlib.Path(charm).name)
-    shutil.copy(charm, fault_charm)
+    tmp_folder = Path("tmp")
+    tmp_folder.mkdir(exist_ok=True)
+    tmp_folder_charm = Path(tmp_folder, charm).absolute()
 
-    logger.info("Inject dependency fault")
-    await inject_dependency_fault(ops_test, MYSQL_APP_NAME, fault_charm)
+    shutil.copy(charm, tmp_folder_charm)
 
-    application = ops_test.model.applications[MYSQL_APP_NAME]
+    logging.info("Inject dependency fault")
+    inject_dependency_fault(juju, MYSQL_APP_NAME, tmp_folder_charm)
 
-    logger.info("Refresh the charm")
-    await application.refresh(path=fault_charm)
+    logging.info("Refresh the charm")
+    juju.refresh(app=MYSQL_APP_NAME, path=tmp_folder_charm)
 
-    logger.info("Get first upgrading unit")
-    # highest ordinal unit always the first to upgrade
-    unit = get_unit_by_number(MYSQL_APP_NAME, application.units, 2)
-
-    logger.info("Wait for upgrade to fail on first upgrading unit")
-    await ops_test.model.block_until(
-        lambda: unit.workload_status == "blocked",
-        timeout=TIMEOUT,
+    logging.info("Wait for upgrade to fail on first upgrading unit")
+    juju.wait(
+        ready=wait_for_unit_status(MYSQL_APP_NAME, mysql_upgrade_unit, "blocked"),
+        timeout=10 * MINUTE_SECS,
     )
 
-    logger.info("Ensure continuous_writes while in failure state on remaining units")
-    mysql_units = [unit_ for unit_ in application.units if unit_.name != unit.name]
-    await ensure_all_units_continuous_writes_incrementing(
-        ops_test, credentials=credentials, mysql_units=mysql_units
+    logging.info("Ensure continuous writes on remaining units")
+    mysql_remaining_units = [unit for unit in mysql_app_units if unit != mysql_upgrade_unit]
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME, mysql_remaining_units)
+
+    logging.info("Re-run pre-upgrade-check action")
+    task = juju.run(unit=mysql_app_leader, action="pre-upgrade-check")
+    task.raise_on_failure()
+
+    logging.info("Re-refresh the charm")
+    juju.refresh(app=MYSQL_APP_NAME, path=charm)
+
+    logging.info("Wait for upgrade to complete on first upgrading unit")
+    juju.wait(
+        ready=wait_for_unit_message(MYSQL_APP_NAME, mysql_upgrade_unit, "upgrade completed"),
+        timeout=10 * MINUTE_SECS,
     )
 
-    logger.info("Re-run pre-upgrade-check action")
-    await juju_.run_action(leader_unit, "pre-upgrade-check")
+    logging.info("Resume upgrade")
+    while get_k8s_stateful_set_partitions(juju, MYSQL_APP_NAME) == 2:
+        # ignore action return error as it is expected when
+        # the leader unit is the next one to be upgraded
+        # due it being immediately rolled when the partition
+        # is patched in the stateful set
+        with suppress(TaskError):
+            task = juju.run(unit=mysql_app_leader, action="resume-upgrade")
+            task.raise_on_failure()
 
-    logger.info("Re-refresh the charm")
-    await application.refresh(path=charm)
-
-    logger.info("Wait for upgrade to complete on first upgrading unit")
-    await ops_test.model.block_until(
-        lambda: unit.workload_status_message == "upgrade completed", timeout=TIMEOUT
+    logging.info("Wait for upgrade to recover")
+    juju.wait(
+        ready=lambda status: jubilant_backports.all_active(status, MYSQL_APP_NAME),
+        timeout=20 * MINUTE_SECS,
     )
 
-    logger.info("Resume upgrade")
-    while get_sts_partition(ops_test, MYSQL_APP_NAME) == 2:
-        # resume action sometime fails in CI, no clear reason
-        try:
-            await juju_.run_action(leader_unit, "resume-upgrade")
-            sleep(2)
-        except AssertionError:
-            # ignore action return error as it is expected when
-            # the leader unit is the next one to be upgraded
-            # due it being immediately rolled when the partition
-            # is patched in the statefulset
-            pass
+    logging.info("Ensure continuous writes after rollback procedure")
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME, list(mysql_app_units))
 
-    logger.info("Wait for application to recover")
-    await ops_test.model.block_until(
-        lambda: all(unit.workload_status == "active" for unit in application.units),
-        timeout=TIMEOUT,
-    )
-
-    logger.info("Ensure continuous_writes after rollback procedure")
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
-
-    # remove fault charm file
-    fault_charm.unlink()
+    # Remove fault charm file
+    tmp_folder_charm.unlink()
 
 
-async def inject_dependency_fault(
-    ops_test: OpsTest, application_name: str, charm_file: Union[str, Path]
-) -> None:
+def inject_dependency_fault(juju: Juju, app_name: str, charm_file: str | Path) -> None:
     """Inject a dependency fault into the mysql charm."""
     # Open dependency.json and load current charm version
     with open("src/dependency.json") as dependency_file:
         current_charm_version = json.load(dependency_file)["charm"]["version"]
 
-    # query running dependency to overwrite with incompatible version
-    relation_data = await get_relation_data(ops_test, application_name, "upgrade")
+    # Query running dependency to overwrite with incompatible version
+    relation_data = get_relation_data(juju, app_name, "upgrade")
 
     loaded_dependency_dict = json.loads(relation_data[0]["application-data"]["dependencies"])
     loaded_dependency_dict["charm"]["upgrade_supported"] = f">{current_charm_version}"
